Refactor some public app code a bit to fix issues running some tests independently (eg, just tests/web/test_client.py)
