+++ src/auslib/web/public/client.py
@@ -8,7 +8,7 @@ from flask import make_response
 
 from auslib.AUS import FORCE_FALLBACK_MAPPING, FORCE_MAIN_MAPPING
 from auslib.global_state import dbo
-from auslib.web.public.base import AUS, with_transaction
+from auslib.web.public.helpers import AUS, with_transaction
 
 try:
     from urllib import unquote
